:art: refactor to remove circular dependencies
