+++ code/go/0chain.net/blobbercore/handler/handler.go
@@ -4,16 +4,13 @@ package handler
 
 import (
 	"context"
-	"encoding/json"
 	"net/http"
 	"os"
 	"runtime/pprof"
 
+	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/convert"
+
 	"google.golang.org/grpc/metadata"
-
-	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/allocation"
-
-	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/reference"
 
 	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/blobbergrpc"
 
@@ -142,12 +139,8 @@ func AllocationHandler(svc *blobberGRPCService) func(ctx context.Context, r *htt
 			return nil, err
 		}
 
-		return GetAllocationResponseHandler(getAllocationResp), nil
+		return convert.GetAllocationResponseHandler(getAllocationResp), nil
 	}
-}
-
-func GetAllocationResponseHandler(resp *blobbergrpc.GetAllocationResponse) *allocation.Allocation {
-	return GRPCAllocationToAllocation(resp.Allocation)
 }
 
 func FileMetaHandler(svc *blobberGRPCService) func(ctx context.Context, r *http.Request) (interface{}, error) {
@@ -164,19 +157,8 @@ func FileMetaHandler(svc *blobberGRPCService) func(ctx context.Context, r *http.
 			return nil, err
 		}
 
-		return GetFileMetaDataResponseHandler(getFileMetaDataResp), nil
+		return convert.GetFileMetaDataResponseHandler(getFileMetaDataResp), nil
 	}
-}
-
-func GetFileMetaDataResponseHandler(resp *blobbergrpc.GetFileMetaDataResponse) map[string]interface{} {
-	var collaborators []reference.Collaborator
-	for _, c := range resp.Collaborators {
-		collaborators = append(collaborators, *GRPCCollaboratorToCollaborator(c))
-	}
-
-	result := reference.FileRefGRPCToFileRef(resp.MetaData).GetListingData(context.Background())
-	result["collaborators"] = collaborators
-	return result
 }
 
 func CommitMetaTxnHandler(ctx context.Context, r *http.Request) (interface{}, error) {
@@ -214,23 +196,8 @@ func FileStatsHandler(svc *blobberGRPCService) func(ctx context.Context, r *http
 			return nil, err
 		}
 
-		return GetFileStatsResponseHandler(getFileStatsResponse), nil
+		return convert.GetFileStatsResponseHandler(getFileStatsResponse), nil
 	}
-}
-
-func GetFileStatsResponseHandler(resp *blobbergrpc.GetFileStatsResponse) map[string]interface{} {
-	ctx := context.Background()
-	result := reference.FileRefGRPCToFileRef(resp.MetaData).GetListingData(ctx)
-
-	statsMap := make(map[string]interface{})
-	statsBytes, _ := json.Marshal(FileStatsGRPCToFileStats(resp.Stats))
-	_ = json.Unmarshal(statsBytes, &statsMap)
-
-	for k, v := range statsMap {
-		result[k] = v
-	}
-
-	return result
 }
 
 /*DownloadHandler is the handler to respond to download requests from clients*/
@@ -259,21 +226,7 @@ func ListHandler(svc *blobberGRPCService) func(ctx context.Context, r *http.Requ
 			return nil, err
 		}
 
-		return ListEntitesResponseHandler(listEntitiesResponse), nil
-	}
-}
-
-func ListEntitesResponseHandler(resp *blobbergrpc.ListEntitiesResponse) *ListResult {
-	ctx := context.Background()
-	var entities []map[string]interface{}
-	for i := range resp.Entities {
-		entities = append(entities, reference.FileRefGRPCToFileRef(resp.Entities[i]).GetListingData(ctx))
-	}
-
-	return &ListResult{
-		AllocationRoot: resp.AllocationRoot,
-		Meta:           reference.FileRefGRPCToFileRef(resp.MetaData).GetListingData(ctx),
-		Entities:       entities,
+		return convert.ListEntitesResponseHandler(listEntitiesResponse), nil
 	}
 }
 
@@ -301,15 +254,7 @@ func ReferencePathHandler(svc *blobberGRPCService) func(ctx context.Context, r *
 			return nil, err
 		}
 
-		return GetReferencePathResponseHandler(getReferencePathResponse), nil
-	}
-}
-
-func GetReferencePathResponseHandler(getReferencePathResponse *blobbergrpc.GetReferencePathResponse) *ReferencePathResult {
-	var recursionCount int
-	return &ReferencePathResult{
-		ReferencePath: ReferencePathGRPCToReferencePath(&recursionCount, getReferencePathResponse.ReferencePath),
-		LatestWM:      WriteMarkerGRPCToWriteMarker(getReferencePathResponse.LatestWM),
+		return convert.GetReferencePathResponseHandler(getReferencePathResponse), nil
 	}
 }
 
@@ -325,27 +270,7 @@ func ObjectPathHandler(svc *blobberGRPCService) func(ctx context.Context, r *htt
 			return nil, err
 		}
 
-		return GetObjectPathResponseHandler(getObjectPathResponse), nil
-	}
-}
-
-func GetObjectPathResponseHandler(getObjectPathResponse *blobbergrpc.GetObjectPathResponse) *ObjectPathResult {
-	ctx := context.Background()
-	path := reference.FileRefGRPCToFileRef(getObjectPathResponse.ObjectPath.Path).GetListingData(ctx)
-	var pathList []map[string]interface{}
-	for _, pl := range getObjectPathResponse.ObjectPath.PathList {
-		pathList = append(pathList, reference.FileRefGRPCToFileRef(pl).GetListingData(ctx))
-	}
-	path["list"] = pathList
-
-	return &ObjectPathResult{
-		ObjectPath: &reference.ObjectPath{
-			RootHash:     getObjectPathResponse.ObjectPath.RootHash,
-			Meta:         reference.FileRefGRPCToFileRef(getObjectPathResponse.ObjectPath.Meta).GetListingData(ctx),
-			Path:         path,
-			FileBlockNum: getObjectPathResponse.ObjectPath.FileBlockNum,
-		},
-		LatestWM: WriteMarkerGRPCToWriteMarker(getObjectPathResponse.LatestWriteMarker),
+		return convert.GetObjectPathResponseHandler(getObjectPathResponse), nil
 	}
 }
 
@@ -360,15 +285,7 @@ func ObjectTreeHandler(svc *blobberGRPCService) func(ctx context.Context, r *htt
 			return nil, err
 		}
 
-		return GetObjectTreeResponseHandler(getObjectTreeResponse), nil
-	}
-}
-
-func GetObjectTreeResponseHandler(getObjectTreeResponse *blobbergrpc.GetObjectTreeResponse) *ReferencePathResult {
-	var recursionCount int
-	return &ReferencePathResult{
-		ReferencePath: ReferencePathGRPCToReferencePath(&recursionCount, getObjectTreeResponse.ReferencePath),
-		LatestWM:      WriteMarkerGRPCToWriteMarker(getObjectTreeResponse.LatestWM),
+		return convert.GetObjectTreeResponseHandler(getObjectTreeResponse), nil
 	}
 }
