+++ code/go/0chain.net/blobbercore/handler/object_operation_handler.go
@@ -6,6 +6,8 @@ import (
 	"encoding/json"
 	"errors"
 
+	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/blobberHTTP"
+
 	"net/http"
 	"path/filepath"
 	"strconv"
@@ -342,7 +344,7 @@ func (fsh *StorageHandler) DownloadFile(ctx context.Context, r *http.Request) (
 	if latestRM != nil &&
 		latestRM.ReadCounter+(numBlocks) != readMarker.ReadCounter {
 
-		var response = &DownloadResponse{
+		var response = &blobberHTTP.DownloadResponse{
 			Success:      false,
 			LatestRM:     latestRM,
 			Path:         fileref.Path,
@@ -398,7 +400,7 @@ func (fsh *StorageHandler) DownloadFile(ctx context.Context, r *http.Request) (
 			"couldn't save latest read marker: %v", err)
 	}
 
-	var response = &DownloadResponse{}
+	var response = &blobberHTTP.DownloadResponse{}
 	response.Success = true
 	response.LatestRM = readMarker
 	response.Data = respData
@@ -409,7 +411,7 @@ func (fsh *StorageHandler) DownloadFile(ctx context.Context, r *http.Request) (
 	return respData, nil
 }
 
-func (fsh *StorageHandler) CommitWrite(ctx context.Context, r *http.Request) (*CommitResult, error) {
+func (fsh *StorageHandler) CommitWrite(ctx context.Context, r *http.Request) (*blobberHTTP.CommitResult, error) {
 
 	if r.Method == "GET" {
 		return nil, common.NewError("invalid_method", "Invalid method used for the upload URL. Use POST instead")
@@ -488,7 +490,7 @@ func (fsh *StorageHandler) CommitWrite(ctx context.Context, r *http.Request) (*C
 			err)
 	}
 
-	var result CommitResult
+	var result blobberHTTP.CommitResult
 	var latestWM *writemarker.WriteMarkerEntity
 	if len(allocationObj.AllocationRoot) == 0 {
 		latestWM = nil
@@ -654,7 +656,7 @@ func (fsh *StorageHandler) RenameObject(ctx context.Context, r *http.Request) (i
 		return nil, common.NewError("connection_write_error", "Error writing the connection meta data")
 	}
 
-	result := &UploadResult{}
+	result := &blobberHTTP.UploadResult{}
 	result.Filename = new_name
 	result.Hash = objectRef.Hash
 	result.MerkleRoot = objectRef.MerkleRoot
@@ -853,7 +855,7 @@ func (fsh *StorageHandler) CopyObject(ctx context.Context, r *http.Request) (int
 		return nil, common.NewError("connection_write_error", "Error writing the connection meta data")
 	}
 
-	result := &UploadResult{}
+	result := &blobberHTTP.UploadResult{}
 	result.Filename = objectRef.Name
 	result.Hash = objectRef.Hash
 	result.MerkleRoot = objectRef.MerkleRoot
@@ -862,7 +864,7 @@ func (fsh *StorageHandler) CopyObject(ctx context.Context, r *http.Request) (int
 	return result, nil
 }
 
-func (fsh *StorageHandler) DeleteFile(ctx context.Context, r *http.Request, connectionObj *allocation.AllocationChangeCollector) (*UploadResult, error) {
+func (fsh *StorageHandler) DeleteFile(ctx context.Context, r *http.Request, connectionObj *allocation.AllocationChangeCollector) (*blobberHTTP.UploadResult, error) {
 	path := r.FormValue("path")
 	if len(path) == 0 {
 		return nil, common.NewError("invalid_parameters", "Invalid path")
@@ -884,7 +886,7 @@ func (fsh *StorageHandler) DeleteFile(ctx context.Context, r *http.Request, conn
 		connectionObj.Size += allocationChange.Size
 		connectionObj.AddChange(allocationChange, dfc)
 
-		result := &UploadResult{}
+		result := &blobberHTTP.UploadResult{}
 		result.Filename = fileRef.Name
 		result.Hash = fileRef.Hash
 		result.MerkleRoot = fileRef.MerkleRoot
@@ -897,7 +899,7 @@ func (fsh *StorageHandler) DeleteFile(ctx context.Context, r *http.Request, conn
 }
 
 //WriteFile stores the file into the blobber files system from the HTTP request
-func (fsh *StorageHandler) WriteFile(ctx context.Context, r *http.Request) (*UploadResult, error) {
+func (fsh *StorageHandler) WriteFile(ctx context.Context, r *http.Request) (*blobberHTTP.UploadResult, error) {
 
 	if r.Method == "GET" {
 		return nil, common.NewError("invalid_method", "Invalid method used for the upload URL. Use multi-part form POST / PUT / DELETE instead")
@@ -941,7 +943,7 @@ func (fsh *StorageHandler) WriteFile(ctx context.Context, r *http.Request) (*Upl
 	mutex.Lock()
 	defer mutex.Unlock()
 
-	result := &UploadResult{}
+	result := &blobberHTTP.UploadResult{}
 	mode := allocation.INSERT_OPERATION
 	if r.Method == "PUT" {
 		mode = allocation.UPDATE_OPERATION
