+++ code/go/0chain.net/blobbercore/reference/objectpath.go
@@ -3,9 +3,6 @@ package reference
 import (
 	"context"
 	"fmt"
-	"time"
-
-	"github.com/0chain/blobber/code/go/0chain.net/blobbercore/blobbergrpc"
 
 	"github.com/0chain/blobber/code/go/0chain.net/core/common"
 )
@@ -89,144 +86,4 @@ func GetObjectPath(ctx context.Context, allocationID string, blockNum int64) (*O
 	retObj.RefID = curRef.ID
 
 	return &retObj, nil
-}
-
-func FileRefToFileRefGRPC(ref *Ref) *blobbergrpc.FileRef {
-	if ref == nil {
-		return nil
-	}
-
-	var fileMetaData *blobbergrpc.FileMetaData
-	var dirMetaData *blobbergrpc.DirMetaData
-	switch ref.Type {
-	case FILE:
-		fileMetaData = convertFileRefToFileMetaDataGRPC(ref)
-	case DIRECTORY:
-		dirMetaData = convertDirRefToDirMetaDataGRPC(ref)
-	}
-
-	return &blobbergrpc.FileRef{
-		Type:         ref.Type,
-		FileMetaData: fileMetaData,
-		DirMetaData:  dirMetaData,
-	}
-}
-
-func convertFileRefToFileMetaDataGRPC(fileref *Ref) *blobbergrpc.FileMetaData {
-	var commitMetaTxnsGRPC []*blobbergrpc.CommitMetaTxn
-	for _, c := range fileref.CommitMetaTxns {
-		commitMetaTxnsGRPC = append(commitMetaTxnsGRPC, &blobbergrpc.CommitMetaTxn{
-			RefId:     c.RefID,
-			TxnId:     c.TxnID,
-			CreatedAt: c.CreatedAt.UnixNano(),
-		})
-	}
-	return &blobbergrpc.FileMetaData{
-		Type:                fileref.Type,
-		LookupHash:          fileref.LookupHash,
-		Name:                fileref.Name,
-		Path:                fileref.Path,
-		Hash:                fileref.Hash,
-		NumBlocks:           fileref.NumBlocks,
-		PathHash:            fileref.PathHash,
-		CustomMeta:          fileref.CustomMeta,
-		ContentHash:         fileref.ContentHash,
-		Size:                fileref.Size,
-		MerkleRoot:          fileref.MerkleRoot,
-		ActualFileSize:      fileref.ActualFileSize,
-		ActualFileHash:      fileref.ActualFileHash,
-		MimeType:            fileref.MimeType,
-		ThumbnailSize:       fileref.ThumbnailSize,
-		ThumbnailHash:       fileref.ThumbnailHash,
-		ActualThumbnailSize: fileref.ActualThumbnailSize,
-		ActualThumbnailHash: fileref.ActualThumbnailHash,
-		EncryptedKey:        fileref.EncryptedKey,
-		Attributes:          fileref.Attributes,
-		OnCloud:             fileref.OnCloud,
-		CommitMetaTxns:      commitMetaTxnsGRPC,
-		CreatedAt:           fileref.CreatedAt.UnixNano(),
-		UpdatedAt:           fileref.UpdatedAt.UnixNano(),
-	}
-}
-
-func convertDirRefToDirMetaDataGRPC(dirref *Ref) *blobbergrpc.DirMetaData {
-	return &blobbergrpc.DirMetaData{
-		Type:       dirref.Type,
-		LookupHash: dirref.LookupHash,
-		Name:       dirref.Name,
-		Path:       dirref.Path,
-		Hash:       dirref.Hash,
-		NumBlocks:  dirref.NumBlocks,
-		PathHash:   dirref.PathHash,
-		Size:       dirref.Size,
-		CreatedAt:  dirref.CreatedAt.UnixNano(),
-		UpdatedAt:  dirref.UpdatedAt.UnixNano(),
-	}
-}
-
-func FileRefGRPCToFileRef(ref *blobbergrpc.FileRef) *Ref {
-	if ref == nil {
-		return nil
-	}
-
-	switch ref.Type {
-	case FILE:
-		return convertFileMetaDataGRPCToFileRef(ref.FileMetaData)
-	case DIRECTORY:
-		return convertDirMetaDataGRPCToDirRef(ref.DirMetaData)
-	}
-
-	return nil
-}
-
-func convertFileMetaDataGRPCToFileRef(metaData *blobbergrpc.FileMetaData) *Ref {
-	var commitMetaTxnsGRPC []CommitMetaTxn
-	for _, c := range metaData.CommitMetaTxns {
-		commitMetaTxnsGRPC = append(commitMetaTxnsGRPC, CommitMetaTxn{
-			RefID:     c.RefId,
-			TxnID:     c.TxnId,
-			CreatedAt: time.Unix(0, c.CreatedAt),
-		})
-	}
-	return &Ref{
-		Type:                metaData.Type,
-		LookupHash:          metaData.LookupHash,
-		Name:                metaData.Name,
-		Path:                metaData.Path,
-		Hash:                metaData.Hash,
-		NumBlocks:           metaData.NumBlocks,
-		PathHash:            metaData.PathHash,
-		CustomMeta:          metaData.CustomMeta,
-		ContentHash:         metaData.ContentHash,
-		Size:                metaData.Size,
-		MerkleRoot:          metaData.MerkleRoot,
-		ActualFileSize:      metaData.ActualFileSize,
-		ActualFileHash:      metaData.ActualFileHash,
-		MimeType:            metaData.MimeType,
-		ThumbnailSize:       metaData.ThumbnailSize,
-		ThumbnailHash:       metaData.ThumbnailHash,
-		ActualThumbnailSize: metaData.ActualThumbnailSize,
-		ActualThumbnailHash: metaData.ActualThumbnailHash,
-		EncryptedKey:        metaData.EncryptedKey,
-		Attributes:          metaData.Attributes,
-		OnCloud:             metaData.OnCloud,
-		CommitMetaTxns:      commitMetaTxnsGRPC,
-		CreatedAt:           time.Unix(0, metaData.CreatedAt),
-		UpdatedAt:           time.Unix(0, metaData.UpdatedAt),
-	}
-}
-
-func convertDirMetaDataGRPCToDirRef(dirref *blobbergrpc.DirMetaData) *Ref {
-	return &Ref{
-		Type:       dirref.Type,
-		LookupHash: dirref.LookupHash,
-		Name:       dirref.Name,
-		Path:       dirref.Path,
-		Hash:       dirref.Hash,
-		NumBlocks:  dirref.NumBlocks,
-		PathHash:   dirref.PathHash,
-		Size:       dirref.Size,
-		CreatedAt:  time.Unix(0, dirref.CreatedAt),
-		UpdatedAt:  time.Unix(0, dirref.UpdatedAt),
-	}
 }
